+++ NEWS.md
@@ -2,8 +2,10 @@
 
 * Initial complete release
 
-# shadow 0.3.1
+# shadow 0.3.2
 
 * Fixed 'toSeg' behavior when 'x' has single attribute 
+* Fixed 'shiftAz' behavior when 'object' is 'SpatialPoints*' 
+* 'parallel' support in functions 'shadowHeight' and 'SVF'